blogging_system/requirements.txt
appdirs @ file:///home/bibekg/.cache/pypoetry/artifacts/47/cf/4f/4ef02fb715aa36daeebad18cc5570126159c659c41c7b5ec46a7387d9b/appdirs-1.4.4-py2.py3-none-any.whl
asgiref @ file:///home/bibekg/.cache/pypoetry/artifacts/9c/90/26/d9fa1dfd567d8ba46faa44b741eb6442f3b97eb9f10a40bc1ad7a7f10e/asgiref-3.3.1-py3-none-any.whl
attrs @ file:///home/bibekg/.cache/pypoetry/artifacts/f9/48/82/553e4bef24d3b294c0c18f27a7853f3ed151508efd144cb7ea37db1c48/attrs-20.3.0-py2.py3-none-any.whl
black @ file:///home/bibekg/.cache/pypoetry/artifacts/56/00/d0/a3bee2cbd161f599bac7d8ee24fb2b17aeba39b53bc42724de21f352bf/black-20.8b1.tar.gz
click @ file:///home/bibekg/.cache/pypoetry/artifacts/e2/79/34/a23e9d2f683ed66be11ec3bd760dec3a2fe228cfdedf2071bcf0531b06/click-7.1.2-py2.py3-none-any.whl
Django @ file:///home/bibekg/.cache/pypoetry/artifacts/80/9f/e8/7dbac946baeff06413dea73548571fa188758ff084c6258f4e50a09137/Django-3.1.7-py3-none-any.whl
django-ckeditor @ file:///home/bibekg/.cache/pypoetry/artifacts/71/6b/ff/161e183b25f8fd2bef886b709d80ebbd08f8a35093b07653d8a792580c/django_ckeditor-6.0.0-py2.py3-none-any.whl
django-js-asset @ file:///home/bibekg/.cache/pypoetry/artifacts/00/7c/b4/e3ecaf08b9600fd79e9d915c0d908b307b9ce8de512033d71731d19cbc/django_js_asset-1.2.2-py2.py3-none-any.whl
django-schema-graph @ file:///home/bibekg/.cache/pypoetry/artifacts/fc/b4/10/01c868d54a1538c0db4e0863fa5a70dc72347d1a376b1bc718cd179dcf/django_schema_graph-1.2.0-py2.py3-none-any.whl
django-taggit @ file:///home/bibekg/.cache/pypoetry/artifacts/b5/67/84/317a22d63486210e420ef2d45a6206f7fe1696ba9cf3ba75b6fb2c712f/django_taggit-1.3.0-py3-none-any.whl
mypy-extensions @ file:///home/bibekg/.cache/pypoetry/artifacts/92/45/bf/1807ce854ff668d92602207a37bfa9316def2a3f257bd03c4c5be4bc9b/mypy_extensions-0.4.3-py2.py3-none-any.whl
pathspec @ file:///home/bibekg/.cache/pypoetry/artifacts/40/8b/b2/80a6945971d8475bcc04d09afec4845855dd74f68da6a4c18bbf8f7784/pathspec-0.8.1-py2.py3-none-any.whl
Pillow @ file:///home/bibekg/.cache/pypoetry/artifacts/9d/54/2c/7a7d9eaffb94019fa6fb3d10973cadcb3a8d076da0cb8d175bc90ca197/Pillow-8.1.2-cp39-cp39-manylinux1_x86_64.whl
pytz @ file:///home/bibekg/.cache/pypoetry/artifacts/b0/a7/8d/54de3ab4d1ff29abbbca1e9ccbaefdc2a1b290138311b84f73bee16de1/pytz-2021.1-py2.py3-none-any.whl
regex @ file:///home/bibekg/.cache/pypoetry/artifacts/a5/0c/c1/33400529698ff03037d95e34dd929002b124500316b658d69cd202001e/regex-2020.11.13-cp39-cp39-manylinux2014_x86_64.whl
six @ file:///home/bibekg/.cache/pypoetry/artifacts/dd/1c/65/ad0dea11136f5a869f072890a0eea955aa8fc35b90c85c55249fd3abfe/six-1.15.0-py2.py3-none-any.whl
sqlparse @ file:///home/bibekg/.cache/pypoetry/artifacts/bc/af/02/8aa71b17fd39b26b858293918c5c6412f83768a0bfd0a4626f660aa5f8/sqlparse-0.4.1-py3-none-any.whl
toml @ file:///home/bibekg/.cache/pypoetry/artifacts/6b/6a/c9/53b19f7870a77d855e8b05ecdc98193944e5d246dafe11bbcad850ecba/toml-0.10.2-py2.py3-none-any.whl
typed-ast @ file:///home/bibekg/.cache/pypoetry/artifacts/cd/79/52/46fe8f18205909c174875a57775caf1386309766aad7e887cc0d4b023d/typed_ast-1.4.2-cp39-cp39-manylinux1_x86_64.whl
typing-extensions @ file:///home/bibekg/.cache/pypoetry/artifacts/ab/c3/72/446cb2c521f10fc837619e8a7c68ed3c3bd74859bd625b7d74f38a159b/typing_extensions-3.7.4.3-py3-none-any.whl
